@@ -27,8 +27,9 @@ module.exports = function (config) {
     port: 9876,
     colors: true,
     logLevel: config.LOG_INFO,
+    browserNoActivityTimeout: 60000,
     autoWatch: true,
     browsers: ['PhantomJS', 'Chrome'],
-    singleRun: true
+    singleRun: false
   });
 };
