@@ -1,25 +1,39 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+// import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { HomeComponent } from './home.component';
+// import { HomeComponent } from './home.component';
+// import { RootComponent } from '../root/root.component';
+// import { NavbarComponent } from '../navbar/navbar.component';
+// import { RequestService } from '../../services/request-service/request.service';
+// import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+// import { RouterTestingModule } from '@angular/router/testing';
+// import { AuthenticationService } from '../../services/authentication-service/authentication.service';
+// import { SkillsetComponent } from '../skillset/skillset.component';
+// import { ClientMappedComponent } from '../client-mapped/client-mapped.component';
 
-describe('HomeComponent', () => {
-  let component: HomeComponent;
-  let fixture: ComponentFixture<HomeComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ HomeComponent ]
-    })
-    .compileComponents();
-  }));
+// describe('HomeComponent', () => {
+//   let component: HomeComponent;
+//   let fixture: ComponentFixture<HomeComponent>;
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(HomeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(async(() => {
+//     TestBed.configureTestingModule({
+//       declarations: [ HomeComponent, RootComponent, NavbarComponent, HomeComponent, SkillsetComponent, ClientMappedComponent ],
+//       imports: [
+//         HttpClientTestingModule,
+//         RouterTestingModule
+//       ],
+//       providers: [ RequestService, AuthenticationService ]
+//     })
+//     .compileComponents();
+//   }));
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(HomeComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
+
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });
