@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { RouterLink } from '@angular/router';
+import { Router } from '@angular/router';
+import { AuthenticationService } from '../../services/authentication-service/authentication.service';
+import { User } from '../../models/user.model';
 
 @Component({
   selector: 'app-navbar',
@@ -8,9 +10,33 @@ import { RouterLink } from '@angular/router';
 })
 export class NavbarComponent implements OnInit {
 
-  constructor() { }
+  user: User;
+  isAdmin: boolean;
+  isAssociate: boolean;
+
+  constructor(private router: Router, private authService: AuthenticationService) { }
 
   ngOnInit() {
+    this.user = this.authService.getUser();
+
+    if(this.user.tfRoleId === 1){
+      this.isAdmin = true;
+    } else {
+      this.isAdmin = false;
+    }
+
+    if(this.user.tfRoleId === 4){
+      this.isAssociate = true;
+    } else {
+      this.isAssociate = false;
+    }
+
+
+  }
+
+  logout(){
+    this.authService.logout();
+    this.router.navigateByUrl('/login');
   }
 
 }
