@@ -6,10 +6,10 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./root.component.css']
 })
 export class RootComponent implements OnInit {
+  title = 'app-root';
 
   constructor() { }
 
   ngOnInit() {
   }
-
 }
