@@ -1,12 +1,20 @@
 package com.revature.dao;
 
+import java.io.IOException;
 import java.math.BigDecimal;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeSet;
 
 import javax.persistence.NoResultException;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 
+import org.hibernate.Criteria;
 import org.hibernate.Hibernate;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -16,91 +24,84 @@ import org.hibernate.query.Query;
 import com.revature.entity.TfAssociate;
 import com.revature.entity.TfClient;
 import com.revature.entity.TfMarketingStatus;
+import com.revature.model.AssociateInfo;
+import com.revature.model.ClientInfo;
+import com.revature.model.MarketingStatusInfo;
+import com.revature.utils.Dao2DoMapper;
 import com.revature.utils.HibernateUtil;
 import com.revature.utils.LogUtil;
 
 public class AssociateDaoHibernate implements AssociateDao {
 
-    /**
-     * Get a associate from the database given its id.
-     * 
-     * @param associateid
-     */
-    @Override
-    public TfAssociate getAssociate(BigDecimal associateid) {
-        TfAssociate associate;
-        SessionFactory sessionFactory = HibernateUtil.getSession();
-        try (Session session = sessionFactory.openSession()) {
-            CriteriaBuilder builder = session.getCriteriaBuilder();
-            CriteriaQuery<TfAssociate> criteriaQuery = builder.createQuery(TfAssociate.class);
-            Root<TfAssociate> root = criteriaQuery.from(TfAssociate.class);
-            criteriaQuery.select(root).where(builder.equal(root.get("tfAssociateId"), associateid));
-            Query<TfAssociate> query = session.createQuery(criteriaQuery);
-            try {
-                associate = query.getSingleResult();
+	/**
+	 * Get a associate from the database given its id.
+	 * 
+	 * @param associateid
+	 * @throws IOException 
+	 */
+	@Override
+	public AssociateInfo getAssociate(BigDecimal associateid, Session session) throws IOException {
+		TfAssociate associate;
+		
+		CriteriaBuilder builder = session.getCriteriaBuilder();
+		CriteriaQuery<TfAssociate> criteriaQuery = builder.createQuery(TfAssociate.class);
+		Root<TfAssociate> root = criteriaQuery.from(TfAssociate.class);
+		criteriaQuery.select(root).where(builder.equal(root.get("tfAssociateId"), associateid));
+		Query<TfAssociate> query = session.createQuery(criteriaQuery);
 
-                Hibernate.initialize(associate.getTfMarketingStatus());
-                Hibernate.initialize(associate.getTfClient());
-                Hibernate.initialize(associate.getTfEndClient());
-                Hibernate.initialize(associate.getTfBatch());
-            } catch (NoResultException nre) {
-            	LogUtil.logger.error(nre);
-                associate = new TfAssociate();
-            }
-        }
-        return associate;
-    }
+			associate = query.getSingleResult();
+		
+		return Dao2DoMapper.map(associate);
+	}
 
-    /**
-     * Updates an associate's marketing status and client in the database.
-     * 
-     * @param id
-     *            - The ID of the associate to update.
-     * @param marketingStatus
-     *            - A TfMarketingStatus object with the status to change the
-     *            associate to.
-     * @param client
-     *            - A TfClient object with what client the associate will be mapped
-     *            to.
-     */
-    @Override
-    public void updateInfo(BigDecimal id, TfMarketingStatus marketingStatus, TfClient client) {
+	/**
+	 * Updates an associate's marketing status and client in the database.
+	 * 
+	 * @param id
+	 *            - The ID of the associate to update.
+	 * @param marketingStatus
+	 *            - A TfMarketingStatus object with the status to change the
+	 *            associate to.
+	 * @param client
+	 *            - A TfClient object with what client the associate will be mapped
+	 *            to.
+	 * @throws IOException 
+	 */
+	@Override
+	public void updateInfo(Session session, BigDecimal id, MarketingStatusInfo marketingStatus, ClientInfo client) throws IOException {
 
-        SessionFactory factory = HibernateUtil.getSession();
-        try (Session session = factory.openSession()) {
+		TfClient tfclient = null;
+		if (client.getId() != null) {
+			tfclient = session.get(TfClient.class, client.getId());
+		}
+		
+		TfMarketingStatus tfms = session.get(TfMarketingStatus.class, marketingStatus.getId());
 
-            TfMarketingStatus status = null;
-            if (marketingStatus.getTfMarketingStatusId() != null) {
-                status = session.get(TfMarketingStatus.class, marketingStatus.getTfMarketingStatusId());
-            }
+		TfAssociate associate = session.load(TfAssociate.class, id);
+		associate.setTfMarketingStatus(tfms);
+		associate.setTfClient(tfclient);
+		session.saveOrUpdate(associate);
 
-            TfClient tfclient = null;
-            if (client.getTfClientId() != null) {
-                tfclient = session.get(TfClient.class, client.getTfClientId());
-            }
+	}
 
-            Transaction transaction = null;
-            try {
-                transaction = session.beginTransaction();
-                TfAssociate associate = session.load(TfAssociate.class, id);
-                associate.setTfMarketingStatus(status);
-                associate.setTfClient(tfclient);
-
-                session.saveOrUpdate(associate);
-
-                transaction.commit();
-                
-                //clear associates list to force update to stored list(s)
-                HomeDaoImpl.clearAssociates();
-            } catch (Exception e) {
-            	LogUtil.logger.error(e);
-                if (transaction != null) {
-                    transaction.rollback();
-                }
-
-            } finally {
-                session.close();
-            }
-        }
-    }
+	@Override
+	public Map<BigDecimal, AssociateInfo> getAssociates(Session session) {
+		List<TfAssociate> associatesEnt;
+		Map<BigDecimal, AssociateInfo> map = new HashMap<>();
+		CriteriaBuilder cb = session.getCriteriaBuilder();
+		CriteriaQuery<TfAssociate> cq = cb.createQuery(TfAssociate.class);
+		Root<TfAssociate> from = cq.from(TfAssociate.class);
+		CriteriaQuery<TfAssociate> all = cq.select(from);
+		Query<TfAssociate> tq = session.createQuery(all);
+		
+		associatesEnt = tq.getResultList();
+		if(associatesEnt != null) {
+			for(TfAssociate tfa : associatesEnt) {
+				map.put(tfa.getTfAssociateId(), Dao2DoMapper.map(tfa));
+				AssociateInfo.appendToMap(tfa.getTfMarketingStatus());
+			}
+		}
+		
+		return map;
+	}
 }
