@@ -1,6 +1,12 @@
 package com.revature.dao;
 
+import java.io.IOException;
+import java.math.BigDecimal;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import javax.persistence.NoResultException;
 import javax.persistence.criteria.CriteriaBuilder;
@@ -8,17 +14,20 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 
 import org.hibernate.Hibernate;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 import org.hibernate.query.Query;
 
 import com.revature.entity.TfClient;
+import com.revature.model.ClientInfo;
+import com.revature.model.CurriculumInfo;
+import com.revature.utils.Dao2DoMapper;
 import com.revature.utils.HibernateUtil;
 import com.revature.utils.LogUtil;
 
 public class ClientDaoImpl implements ClientDao {
-
-	private static List<TfClient> clients;
 
 	/**
 	 * Get information about a singular client.
@@ -26,55 +35,42 @@ public class ClientDaoImpl implements ClientDao {
 	 * @param name
 	 *            - The name of the client to retrieve.
 	 * @return - A TfClient object with information about the client.
+	 * @throws IOException
 	 */
 	@Override
-	public TfClient getClient(String name) {
+	public TfClient getClient(Session session, String name) throws IOException {
 		TfClient client;
-		SessionFactory sessionFactory = HibernateUtil.getSession();
-		try (Session session = sessionFactory.openSession()) {
 
-			CriteriaBuilder builder = session.getCriteriaBuilder();
-			CriteriaQuery<TfClient> criteriaQuery = builder.createQuery(TfClient.class);
+		CriteriaBuilder builder = session.getCriteriaBuilder();
+		CriteriaQuery<TfClient> criteriaQuery = builder.createQuery(TfClient.class);
 
-			Root<TfClient> root = criteriaQuery.from(TfClient.class);
+		Root<TfClient> root = criteriaQuery.from(TfClient.class);
 
-			criteriaQuery.select(root).where(builder.equal(root.get("tfClientName"), name));
+		criteriaQuery.select(root).where(builder.equal(root.get("tfClientName"), name));
 
-			Query<TfClient> query = session.createQuery(criteriaQuery);
+		Query<TfClient> query = session.createQuery(criteriaQuery);
 
-			try {
-				client = query.getSingleResult();
-			} catch (NoResultException nre) {
-				client = new TfClient();
-				LogUtil.logger.error(nre);
-			}
+		try {
+			client = query.getSingleResult();
+		} catch (NoResultException nre) {
+			client = new TfClient();
+			LogUtil.logger.error(nre);
 		}
 		return client;
 	}
 
 	@Override
-	public List<TfClient> getAllTfClients() {
-		if (clients == null || clients.isEmpty()) {
-			try (Session session = HibernateUtil.getSession().openSession()) {
-				CriteriaQuery<TfClient> cq = session.getCriteriaBuilder().createQuery(TfClient.class);
-				cq.from(TfClient.class);
-				clients = session.createQuery(cq).getResultList();
-				for (TfClient client : clients) {
-					Hibernate.initialize(client.getTfClientId());
-					Hibernate.initialize(client.getTfClientName());
-					Hibernate.initialize(client.getTfAssociates());
-				}
-			}
-		}
-		return clients;
-	}
+	public Map<BigDecimal, ClientInfo> getAllTfClients(Session session) throws HibernateException, IOException {
 
-	/**
-	 * Clears clients list in ClientDaoImpl class.
-	 */
-	public void clearClients() {
-		if (clients != null) {
-			clients.clear();
+		CriteriaQuery<TfClient> cq = session.getCriteriaBuilder().createQuery(TfClient.class);
+		cq.from(TfClient.class);
+		List<TfClient> clients = session.createQuery(cq).getResultList();
+		Map<BigDecimal, ClientInfo> map = new HashMap<>();
+		
+		for (TfClient client : clients) {
+			map.put(client.getTfClientId(), Dao2DoMapper.map(client));
 		}
+
+		return map;
 	}
 }
