@@ -1,40 +1,43 @@
 package com.revature.dao;
 
+import java.io.IOException;
+
 import javax.persistence.StoredProcedureQuery;
 
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 
 import com.revature.utils.HibernateUtil;
 import com.revature.utils.LogUtil;
 
 public class DatabaseDAOImpl {
-	public String deleteAll() {
-		return runStoredProcedure("admin.truncateAllDevTeam", "Database Emptied Successfully", "Database Empty Error");
+	public String deleteAll(Session session) throws HibernateException, IOException {
+		return runStoredProcedure("admin.truncateAllDevTeam", "Database Emptied Successfully", "Database Empty Error",
+				session);
 	}
 
-	public String populate() {
+	public String populate(Session session) throws HibernateException, IOException {
 		return runStoredProcedure("admin.populateAllTables_PROC", "Database Population Successful",
-				"Error: Data Exists");
+				"Error: Data Exists", session);
 	}
 
-	public String populateSF() {
+	public String populateSF(Session session) throws HibernateException, IOException {
 		return runStoredProcedure("admin.populateAllTablesSF_PROC", "SF - Database Population Successful",
-				"Error: Data Exists");
+				"Error: Data Exists", session);
 	}
 
-	private String runStoredProcedure(String queryName, String successMessage, String errorMessage) {
-		try (Session session = HibernateUtil.getSession().openSession()) {
-			String message;
-			try {
-				StoredProcedureQuery query = session.createStoredProcedureQuery(queryName);
-				query.execute();
-				message = successMessage;
-				return message;
-			} catch (Exception e) {
-				LogUtil.logger.error(e);
-				message = errorMessage;
-				return message;
-			}
+	private String runStoredProcedure(String queryName, String successMessage, String errorMessage, Session session)
+			throws HibernateException, IOException {
+		String message;
+		try {
+			StoredProcedureQuery query = session.createStoredProcedureQuery(queryName);
+			query.execute();
+			message = successMessage;
+			return message;
+		} catch (Exception e) {
+			LogUtil.logger.error(e);
+			message = errorMessage;
+			return message;
 		}
 	}
 }
