@@ -1,118 +1,135 @@
 package com.revature.entity;
-// Generated Nov 7, 2017 9:24:46 PM by Hibernate Tools 5.2.5.Final
 
 import java.math.BigDecimal;
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 
 /**
  * TfUser generated by hbm2java
  */
 @Entity
-@Table(name = "TF_USER", schema = "ADMIN")
+@Table(name = "TF_USER", schema="ADMIN")
 public class TfUser implements java.io.Serializable {
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 706405344864879997L;
-	private BigDecimal tfUserId;
-	private TfRole tfRole;
-	private String tfUsername;
-	private String tfHashpassword;
 
-	public TfUser() {
-	}
+    private static final long serialVersionUID = 706405344864879997L;
+    @Id 
+    @Column(name = "TF_USER_ID")
+    /* ID's 1-14 are resereved for manual insertion */ 
+    @SequenceGenerator(sequenceName = "UserId_seq", name = "UserIdSeq", initialValue=15)
+    @GeneratedValue(generator = "UserIdSeq", strategy = GenerationType.SEQUENCE)
+    private int tfUserId;
 
-	public TfUser(BigDecimal tfUserId) {
-		this.tfUserId = tfUserId;
-	}
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "TF_ROLE_ID")
+    private TfRole tfRole;
+    
+    @Column(name = "TF_USERNAME", length = 20)
+    private String tfUserUsername;
+    
+    @Column(name = "TF_HASHPASSWORD", length = 200)
+    private String tfHashpassword;
 
-	public TfUser(BigDecimal tfUserId, TfRole tfRole, String tfUserUsername, String tfUserHashpassword) {
-		this.tfUserId = tfUserId;
-		this.tfRole = tfRole;
-		this.tfUsername = tfUserUsername;
-		this.tfHashpassword = tfUserHashpassword;
-	}
+    public TfUser() {
+    }
 
-	@Id
+    public TfUser(int tfUserId) {
+        this.tfUserId = tfUserId;
+    }
 
-	@Column(name = "TF_USER_ID", unique = true, nullable = false, precision = 22, scale = 0)
-	public BigDecimal getTfUserId() {
-		return this.tfUserId;
-	}
+    public TfUser(int tfUserId, TfRole tfRole, String tfUserUsername, String tfUserHashpassword) {
+        this.tfUserId = tfUserId;
+        this.tfRole = tfRole;
+        this.tfUserUsername = tfUserUsername; 
+        this.tfHashpassword = tfUserHashpassword;
+    }
+    
+    public TfUser(BigDecimal role, String username, String password) { 
+        this.tfRole = new TfRole(role);
+        this.tfUserUsername = username; 
+        this.tfHashpassword = password; 
+    }
 
-	public void setTfUserId(BigDecimal tfUserId) {
-		this.tfUserId = tfUserId;
-	}
+    public int getTfUserId() {
+        return this.tfUserId;
+    }
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "TF_ROLE_ID")
-	public TfRole getTfRole() {
-		return this.tfRole;
-	}
+    public void setTfUserId(int tfUserId) {
+        this.tfUserId = tfUserId;
+    }
 
-	public void setTfRole(TfRole tfRole) {
-		this.tfRole = tfRole;
-	}
+    public TfRole getTfRole() {
+        return this.tfRole;
+    }
 
-	@Column(name = "TF_USERNAME", length = 20)
-	public String getTfUserUsername() {
-		return this.tfUsername;
-	}
+    public void setTfRole(TfRole tfRole) {
+        this.tfRole = tfRole;
+    }
 
-	public void setTfUserUsername(String tfUserUsername) {
-		this.tfUsername = tfUserUsername;
-	}
+    
+    public String getTfUserUsername() {
+        return this.tfUserUsername;
+    }
 
-	@Column(name = "TF_HASHPASSWORD", length = 200)
-	public String getTfUserHashpassword() {
-		return this.tfHashpassword;
-	}
+    public void setTfUserUsername(String tfUserUsername) {
+        this.tfUserUsername = tfUserUsername;
+    }
 
-	public void setTfUserHashpassword(String tfUserHashpassword) {
-		this.tfHashpassword = tfUserHashpassword;
-	}
+    public String getTfUserHashpassword() {
+        return this.tfHashpassword;
+    }
 
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((tfHashpassword == null) ? 0 : tfHashpassword.hashCode());
-		result = prime * result + ((tfUserId == null) ? 0 : tfUserId.hashCode());
-		result = prime * result + ((tfUsername == null) ? 0 : tfUsername.hashCode());
-		return result;
-	}
+    public void setTfUserHashpassword(String tfUserHashpassword) {
+        this.tfHashpassword = tfUserHashpassword;
+    }
 
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		TfUser other = (TfUser) obj;
-		if (tfHashpassword == null) {
-			if (other.tfHashpassword != null)
-				return false;
-		} else if (!tfHashpassword.equals(other.tfHashpassword))
-			return false;
-		if (tfUserId == null) {
-			if (other.tfUserId != null)
-				return false;
-		} else if (!tfUserId.equals(other.tfUserId))
-			return false;
-		if (tfUsername == null) {
-			if (other.tfUsername != null)
-				return false;
-		} else if (!tfUsername.equals(other.tfUsername))
-			return false;
-		return true;
-	}
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((tfHashpassword == null) ? 0 : tfHashpassword.hashCode());
+        result = prime * result + ((tfUserUsername == null) ? 0 : tfUserUsername.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        TfUser other = (TfUser) obj;
+        if (tfHashpassword == null) {
+            if (other.tfHashpassword != null) {
+                return false;
+            }
+        } else if (!tfHashpassword.equals(other.tfHashpassword)) {
+            return false;
+        }
+        if (tfUserId != other.tfUserId) {
+            return false;
+        }
+        if (tfUserUsername == null) {
+            if (other.tfUserUsername != null) {
+                return false;
+            }
+        } else if (!tfUserUsername.equals(other.tfUserUsername)) {
+            return false;
+        }
+        return true;
+    }
 }
