@@ -1,5 +1,6 @@
 package com.revature.test;
 
+import java.io.IOException;
 import java.math.BigDecimal;
 
 import org.testng.Assert;
@@ -11,7 +12,7 @@ import com.revature.services.AssociateService;
 public class testAssociateService {
 
 	@Test
-	public void testgetAssociatePositive() {
+	public void testgetAssociatePositive() throws IOException {
 		AssociateService associateservice = new AssociateService();
 		BigDecimal bigdecimal = new BigDecimal(25);
 		AssociateInfo associate = associateservice.getAssociate(bigdecimal);
@@ -19,7 +20,7 @@ public class testAssociateService {
 	}
 
     @Test
-    public void testgetAssociateNegative() {
+    public void testgetAssociateNegative() throws IOException {
         AssociateService associateservice = new AssociateService();
         BigDecimal bigdecimal = new BigDecimal(-25);
         AssociateInfo associate = associateservice.getAssociate(bigdecimal);
@@ -27,8 +28,7 @@ public class testAssociateService {
     }
 
     @Test
-    public void testUpdateAssociatePositive()
-    {
+    public void testUpdateAssociatePositive() throws IOException {
         AssociateService associateService = new AssociateService();
         associateService.updateAssociate("266", "MAPPED: RESERVED", "Corner Bakery");
         
@@ -39,7 +39,7 @@ public class testAssociateService {
     }
     
     @Test
-    public void testUpdateAssociateNegative(){
+    public void testUpdateAssociateNegative() throws IOException {
         AssociateService associateService = new AssociateService();
         associateService.updateAssociate("266", "Placed Tomorrow", "Petsmart");
         
